refactor: move types wip
