+++ src/table/components/TableWrapper.tsx
@@ -15,7 +15,7 @@ import useFocusListener from '../hooks/use-focus-listener';
 import useScrollListener from '../hooks/use-scroll-listener';
 import { handleWrapperKeyDown } from '../utils/handle-key-press';
 import { updateFocus, handleResetFocus, getCellElement } from '../utils/handle-accessibility';
-import { RootProps } from '../../types';
+import { RootProps } from '../types';
 
 export default function TableWrapper(props: RootProps) {
   const {
