+++ src/table/components/withSelections.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useContextSelector, TableContext } from '../context';
 import { getSelectionStyle } from '../utils/styling-utils';
 import { getCellSelectionState, SelectionActions } from '../utils/selections-utils';
-import { CellHOC, CellHOCProps } from '../../types';
+import { CellHOC, CellHOCProps } from '../types';
 
 export default function withSelections(CellComponent: CellHOC) {
   const HOC = (props: CellHOCProps) => {
