+++ src/table/utils/handle-accessibility.ts
@@ -1,6 +1,7 @@
 import { MouseEvent } from 'react';
 import { stardust } from '@nebula.js/stardust';
-import { Announce, Cell, TotalsPosition, CellFocusProps, HandleResetFocusProps } from '../../types';
+import { Announce, Cell, TotalsPosition } from '../../types';
+import { CellFocusProps, HandleResetFocusProps } from '../types';
 
 export const getCellElement = (rootElement: HTMLElement, cellCoord: [number, number]) =>
   rootElement.getElementsByClassName('sn-table-row')[cellCoord[0]]?.getElementsByClassName('sn-table-cell')[
